Move all modules into root namespace.
