+++ setup.cfg
@@ -12,13 +12,13 @@ python-tag = py3
 
 [flake8]
 max-line-length = 100
-exclude = configobj
+exclude = _vendor
 select = E,F,W
 ignore = E123,E126,E203,E226,E241,E704,W503,W504
 
 [pydocstyle]
 match = ^((?!\.sync-zenodo-metadata|setup).)*\.py$
-match-dir = ^((?!\.|tests|configobj).)*$
+match-dir = ^((?!\.|tests|_vendor).)*$
 ignore-decorators = "deprecated"
 add-ignore = D105, D107, D203, D204, D213
 
@@ -31,8 +31,8 @@ concurrency = thread,multiprocessing
 parallel = True
 source = signac
 omit = 
-	*/signac/common/configobj/*.py
-	*/signac/common/deprecation/*.py
+	*/signac/_vendor/configobj/*.py
+	*/signac/_vendor/deprecation/*.py
 
 [tool:pytest]
 xfail_strict = True
