+++ signac/__init__.py
@@ -9,13 +9,13 @@ data spaces, streamlines post-processing and analysis, and makes data
 collectively accessible.
 """
 
-from . import contrib, errors, sync
+from . import errors, sync
 from ._synced_collections.backends.collection_json import (
     BufferedJSONAttrDict as JSONDict,
 )
-from .contrib import Project, TemporaryProject, get_job, get_project, init_project
-from .core.h5store import H5Store, H5StoreManager
 from .diff import diff_jobs
+from .h5store import H5Store, H5StoreManager
+from .project import Project, TemporaryProject, get_job, get_project, init_project
 from .version import __version__
 
 # Alias some properties related to buffering into the signac namespace.
@@ -27,7 +27,6 @@ set_buffer_capacity = JSONDict.set_buffer_capacity
 
 __all__ = [
     "__version__",
-    "contrib",
     "errors",
     "sync",
     "Project",
