+++ signac/_vendor/configobj/LICENSE
