+++ signac/_vendor/configobj/__init__.py
