+++ signac/_vendor/configobj/_version.py
