+++ signac/_vendor/configobj/validate.py
