+++ signac/_vendor/deprecation/LICENSE
