+++ signac/_vendor/deprecation/README
