+++ signac/_vendor/deprecation/__init__.py
