--- signac/contrib/__init__.py
@@ -1,19 +1,0 @@
-# Copyright (c) 2017 The Regents of the University of Michigan
-# All rights reserved.
-# This software is licensed under the BSD 3-Clause License.
-"""Contrib submodule containing Project class and indexing features."""
-
-import logging
-
-from .project import Project, TemporaryProject, get_job, get_project, init_project
-
-logger = logging.getLogger(__name__)
-
-
-__all__ = [
-    "Project",
-    "TemporaryProject",
-    "get_project",
-    "init_project",
-    "get_job",
-]
