--- signac/core/__init__.py
@@ -1,4 +1,0 @@
-# Copyright (c) 2017 The Regents of the University of Michigan
-# All rights reserved.
-# This software is licensed under the BSD 3-Clause License.
-"""Core submodule containing data structures."""
