+++ signac/dict_manager.py
