+++ signac/filterparse.py
