+++ signac/migration/v1_to_v2.py
@@ -16,9 +16,9 @@ This migration involves the following changes:
 import os
 
 from signac._synced_collections.backends.collection_json import BufferedJSONAttrDict
-from signac.common import configobj
-from signac.common.config import _get_project_config_fn
-from signac.contrib.project import Project
+from signac._vendor import configobj
+from signac.config import _get_project_config_fn
+from signac.project import Project
 
 from .v0_to_v1 import _load_config_v1
