+++ signac/schema.py
@@ -9,6 +9,7 @@ from pprint import pformat
 
 from ._searchindexer import _DictPlaceholder
 from .utility import _nested_dicts_to_dotted_keys
+from .version import __version__
 
 
 class _Vividict(dict):
