+++ signac/utility.py
@@ -1,20 +1,31 @@
 # Copyright (c) 2017 The Regents of the University of Michigan
 # All rights reserved.
 # This software is licensed under the BSD 3-Clause License.
-"""Utilities for signac."""
+"""Utility functions."""
 
-import logging
-import os
+import os.path
 import sys
 from collections.abc import Mapping
 from datetime import timedelta
 from time import time
 
-logger = logging.getLogger(__name__)
-
 
 def _print_err(*args, **kwargs):
     print(*args, file=sys.stderr, **kwargs)
+
+
+def _safe_relpath(path):
+    """Attempt to make a relative path, or return the original path.
+
+    This is useful for logging and representing objects, where an absolute path
+    may be very long.
+    """
+    try:
+        return os.path.relpath(path)
+    except ValueError:
+        # Windows cannot find relative paths across drives, so show the
+        # original path instead.
+        return path
 
 
 def _query_yes_no(question, default="yes"):  # pragma: no cover
