+++ tests/test_job.py
@@ -12,11 +12,14 @@ from tempfile import TemporaryDirectory
 
 import pytest
 
-import signac.common.config
-import signac.contrib
-from signac.contrib.errors import JobsCorruptedError
-from signac.contrib.job import Job
-from signac.errors import DestinationExistsError, InvalidKeyError, KeyTypeError
+import signac.config
+from signac.errors import (
+    DestinationExistsError,
+    InvalidKeyError,
+    JobsCorruptedError,
+    KeyTypeError,
+)
+from signac.job import Job
 
 try:
     import h5py  # noqa: F401
@@ -61,7 +64,7 @@ class TestJobBase:
         request.addfinalizer(self._tmp_dir.cleanup)
         self._tmp_pr = os.path.join(self._tmp_dir.name, "pr")
         os.mkdir(self._tmp_pr)
-        self.config = signac.common.config._load_config()
+        self.config = signac.config._load_config()
         self.project = self.project_class.init_project(path=self._tmp_pr)
 
     def tearDown(self):
