+++ tests/test_shell.py
@@ -12,7 +12,7 @@ import pytest
 from test_project import _initialize_v1_project
 
 import signac
-from signac.common import config
+from signac import config
 
 # Skip linked view tests on Windows
 WINDOWS = sys.platform == "win32"
